Right Click Modifier 1.09 (2018-08-06)
Modify the behavior of right clicking in game

You can enable or disable right click targeting, as well as double click 
targeting, in and out of combat. You can also set a value for the time 
between double clicks. This behavior affects units and objects.

Type "/rcm" in game to open the options menu.